Extract ppp into new module
