+++ lib/sanbase/geoip/data.ex
@@ -4,8 +4,6 @@ defmodule Sanbase.Geoip.Data do
   alias Sanbase.Repo
   alias Sanbase.Geoip
 
-  alias Sanbase.Billing.Plan
-
   schema "geoip_data" do
     field(:ip_address, :string)
     field(:is_vpn, :boolean)
@@ -20,32 +18,6 @@ defmodule Sanbase.Geoip.Data do
     geoip_data
     |> cast(attrs, [:ip_address, :is_vpn, :country_name, :country_code])
     |> validate_required([:ip_address, :is_vpn, :country_name, :country_code])
-  end
-
-  def eligible_for_ppp?(ip_address) do
-    case find_or_insert(ip_address) do
-      {:ok, geoip_data} ->
-        # Turkey
-        geoip_data.country_code in ["TR"] and geoip_data.is_vpn == false
-
-      _ ->
-        false
-    end
-  end
-
-  # Purchasing power parity settings.
-  def ppp_settings(geoip_data) do
-    percent_symbol = Plan.country_ppp_plan_map() |> Map.get(geoip_data.country_code)
-    percent_off = Plan.symbol_to_percent_map() |> Map.get(percent_symbol)
-
-    plans = Plan.ppp_plans_for_country(geoip_data.country_code)
-
-    %{
-      is_eligible_for_ppp: true,
-      plans: plans,
-      country: geoip_data.country_name,
-      percent_off: percent_off
-    }
   end
 
   def country_code_by_ip(ip_address) do
